Use readr in ascat.loadData if installed.
The ascat.loadData function can be quite slow on huge files,
since it uses the base read.table. This adds a check for readr,
and uses that instead if installed, falling back on read.table
otherwise. In the latter case it prints a hint that installing
readr would speed things up. The end result should be identical
either way.